requirements.txt
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.16
django-cors-headers==3.13.0
django-filter==21.1
django-location==2.1.1
django-mailbox==4.8.2
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-social-auth==0.7.28
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==19.7.0
mysqlclient==2.1.0rc1
